Documents/dataflow-observatory/simulators/requirements.txt
confluent-kafka==2.4.0
pyyaml>=6.0
faker==24.4.0
